Initial backend setup Add user and movie modules

+++ src/main/java/com/rewind/rewind/movie/dto/MovieCardResponse.java
@@ -1,15 +1,23 @@
 package com.rewind.rewind.movie.dto;
 
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 public class MovieCardResponse {
+
     private Long id;
     private String title;
     private LocalDate releaseDate;
-    private Double rating;
+    private BigDecimal rating;
     private String photoUrl;
 
-    public MovieCardResponse(Long id, String title, LocalDate releaseDate, Double rating, String photoUrl) {
+    public MovieCardResponse(
+            Long id,
+            String title,
+            LocalDate releaseDate,
+            BigDecimal rating,
+            String photoUrl
+    ) {
         this.id = id;
         this.title = title;
         this.releaseDate = releaseDate;
@@ -18,8 +26,12 @@ public class MovieCardResponse {
     }
 
     public Long getId() { return id; }
+
     public String getTitle() { return title; }
+
     public LocalDate getReleaseDate() { return releaseDate; }
-    public Double getRating() { return rating; }
+
+    public BigDecimal getRating() { return rating; }
+
     public String getPhotoUrl() { return photoUrl; }
 }
