+++ src/main/java/com/rewind/rewind/movie/dto/MovieDetailsResponse.java
@@ -1,23 +1,35 @@
 package com.rewind.rewind.movie.dto;
 
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 public class MovieDetailsResponse {
+
     private Long id;
     private String imdbId;
     private String title;
     private Integer durationMinutes;
     private LocalDate releaseDate;
     private String ageRating;
-    private Double rating;
+    private BigDecimal rating;
     private String shortPlot;
     private String photoUrl;
     private String backdropUrl;
     private String trailerUrl;
 
-    public MovieDetailsResponse(Long id, String imdbId, String title, Integer durationMinutes, LocalDate releaseDate,
-                                String ageRating, Double rating, String shortPlot, String photoUrl,
-                                String backdropUrl, String trailerUrl) {
+    public MovieDetailsResponse(
+            Long id,
+            String imdbId,
+            String title,
+            Integer durationMinutes,
+            LocalDate releaseDate,
+            String ageRating,
+            BigDecimal rating,
+            String shortPlot,
+            String photoUrl,
+            String backdropUrl,
+            String trailerUrl
+    ) {
         this.id = id;
         this.imdbId = imdbId;
         this.title = title;
@@ -32,14 +44,24 @@ public class MovieDetailsResponse {
     }
 
     public Long getId() { return id; }
+
     public String getImdbId() { return imdbId; }
+
     public String getTitle() { return title; }
+
     public Integer getDurationMinutes() { return durationMinutes; }
+
     public LocalDate getReleaseDate() { return releaseDate; }
+
     public String getAgeRating() { return ageRating; }
-    public Double getRating() { return rating; }
+
+    public BigDecimal getRating() { return rating; }
+
     public String getShortPlot() { return shortPlot; }
+
     public String getPhotoUrl() { return photoUrl; }
+
     public String getBackdropUrl() { return backdropUrl; }
+
     public String getTrailerUrl() { return trailerUrl; }
 }
